Extract base class CodeActionTest<TVerifier>
