+++ src/Microsoft.CodeAnalysis.Testing/Microsoft.CodeAnalysis.CodeFix.Testing/CodeFixTest`1.cs
@@ -3,11 +3,10 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
-using System.Reflection;
 using System.Runtime.ExceptionServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeActions;
@@ -19,28 +18,26 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.Testing
 {
-    public abstract class CodeFixTest<TVerifier> : AnalyzerTest<TVerifier>
+    public abstract class CodeFixTest<TVerifier> : CodeActionTest<TVerifier>
         where TVerifier : IVerifier, new()
     {
-        /// <summary>
-        /// Gets or sets the index of the code fix to apply.
-        /// </summary>
-        /// <remarks>
-        /// <para>If <see cref="CodeFixIndex"/> and <see cref="CodeFixEquivalenceKey"/> are both specified, the code fix
-        /// test will further verify that the two properties refer to the same code fix.</para>
-        /// </remarks>
-        /// <seealso cref="CodeFixEquivalenceKey"/>
-        public int? CodeFixIndex { get; set; }
+        /// <inheritdoc cref="CodeActionTest{TVerifier}.CodeActionIndex"/>
+        [Obsolete("Use " + nameof(CodeActionIndex) + " instead.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public int? CodeFixIndex
+        {
+            get => CodeActionIndex;
+            set => CodeActionIndex = value;
+        }
 
-        /// <summary>
-        /// Gets or sets the <see cref="CodeAction.EquivalenceKey"/> of the code fix to apply.
-        /// </summary>
-        /// <remarks>
-        /// <para>If <see cref="CodeFixIndex"/> and <see cref="CodeFixEquivalenceKey"/> are both specified, the code fix
-        /// test will further verify that the two properties refer to the same code fix.</para>
-        /// </remarks>
-        /// <seealso cref="CodeFixIndex"/>
-        public string? CodeFixEquivalenceKey { get; set; }
+        /// <inheritdoc cref="CodeActionTest{TVerifier}.CodeActionEquivalenceKey"/>
+        [Obsolete("Use " + nameof(CodeActionEquivalenceKey) + " instead.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public string? CodeFixEquivalenceKey
+        {
+            get => CodeActionEquivalenceKey;
+            set => CodeActionEquivalenceKey = value;
+        }
 
         /// <summary>
         /// Sets the expected output source file for code fix testing.
@@ -147,16 +144,27 @@ namespace Microsoft.CodeAnalysis.Testing
         /// <seealso href="https://github.com/dotnet/roslyn-sdk/issues/147">#147: Figure out Fix All iteration counts by context</seealso>
         public int? NumberOfFixAllInDocumentIterations { get; set; }
 
-        /// <summary>
-        /// Gets or sets the validation mode for code fixes. The default is
-        /// <see cref="CodeFixValidationMode.SemanticStructure"/>.
-        /// </summary>
-        public CodeFixValidationMode CodeFixValidationMode { get; set; } = CodeFixValidationMode.SemanticStructure;
+        /// <inheritdoc cref="CodeActionTest{TVerifier}.CodeActionValidationMode"/>
+        [Obsolete("Use " + nameof(CodeActionValidationMode) + " instead.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public CodeFixValidationMode CodeFixValidationMode
+        {
+            get => CodeActionValidationMode switch
+            {
+                CodeActionValidationMode.None => CodeFixValidationMode.None,
+                CodeActionValidationMode.SemanticStructure => CodeFixValidationMode.SemanticStructure,
+                CodeActionValidationMode.Full => CodeFixValidationMode.Full,
+                _ => throw new InvalidOperationException(),
+            };
 
-        /// <summary>
-        /// Gets the syntax kind enumeration type for the current code fix test.
-        /// </summary>
-        public abstract Type SyntaxKindType { get; }
+            set => CodeActionValidationMode = value switch
+            {
+                CodeFixValidationMode.None => CodeActionValidationMode.None,
+                CodeFixValidationMode.SemanticStructure => CodeActionValidationMode.SemanticStructure,
+                CodeFixValidationMode.Full => CodeActionValidationMode.Full,
+                _ => throw new ArgumentOutOfRangeException(nameof(value)),
+            };
+        }
 
         protected CodeFixTest()
         {
@@ -210,7 +218,7 @@ namespace Microsoft.CodeAnalysis.Testing
             if (CodeFixExpected())
             {
                 await VerifyDiagnosticsAsync(fixedState.Sources.ToArray(), fixedState.AdditionalFiles.ToArray(), fixedState.AdditionalProjects.ToArray(), fixedState.AdditionalReferences.ToArray(), fixedState.ExpectedDiagnostics.ToArray(), Verify.PushContext("Diagnostics of fixed state"), cancellationToken).ConfigureAwait(false);
-                if (CodeFixExpected(BatchFixedState))
+                if (CodeActionExpected(BatchFixedState))
                 {
                     await VerifyDiagnosticsAsync(batchFixedState.Sources.ToArray(), batchFixedState.AdditionalFiles.ToArray(), batchFixedState.AdditionalProjects.ToArray(), batchFixedState.AdditionalReferences.ToArray(), batchFixedState.ExpectedDiagnostics.ToArray(), Verify.PushContext("Diagnostics of batch fixed state"), cancellationToken).ConfigureAwait(false);
                 }
@@ -221,17 +229,8 @@ namespace Microsoft.CodeAnalysis.Testing
 
         private bool CodeFixExpected()
         {
-            return CodeFixExpected(FixedState)
-                || CodeFixExpected(BatchFixedState);
-        }
-
-        private static bool CodeFixExpected(SolutionState state)
-        {
-            return state.InheritanceMode != null
-                || state.MarkupHandling != null
-                || state.Sources.Any()
-                || state.AdditionalFiles.Any()
-                || state.AdditionalFilesFactories.Any();
+            return CodeActionExpected(FixedState)
+                || CodeActionExpected(BatchFixedState);
         }
 
         /// <summary>
@@ -351,7 +350,7 @@ namespace Microsoft.CodeAnalysis.Testing
             var compilerDiagnostics = await GetCompilerDiagnosticsAsync(project, cancellationToken).ConfigureAwait(false);
 
             ExceptionDispatchInfo? iterationCountFailure;
-            (project, iterationCountFailure) = await getFixedProject(analyzers, codeFixProviders, CodeFixIndex, CodeFixEquivalenceKey, project, numberOfIterations, verifier, cancellationToken).ConfigureAwait(false);
+            (project, iterationCountFailure) = await getFixedProject(analyzers, codeFixProviders, CodeActionIndex, CodeActionEquivalenceKey, project, numberOfIterations, verifier, cancellationToken).ConfigureAwait(false);
 
             // After applying all of the code fixes, compare the resulting string to the inputted one
             var updatedDocuments = project.Documents.ToArray();
@@ -382,12 +381,6 @@ namespace Microsoft.CodeAnalysis.Testing
 
             // Validate the iteration counts after validating the content
             iterationCountFailure?.Throw();
-        }
-
-        private static bool HasAnyChange(SolutionState oldState, SolutionState newState)
-        {
-            return !oldState.Sources.SequenceEqual(newState.Sources, SourceFileEqualityComparer.Instance)
-                || !oldState.AdditionalFiles.SequenceEqual(newState.AdditionalFiles, SourceFileEqualityComparer.Instance);
         }
 
         private async Task<(Project project, ExceptionDispatchInfo? iterationCountFailure)> FixEachAnalyzerDiagnosticAsync(ImmutableArray<DiagnosticAnalyzer> analyzers, ImmutableArray<CodeFixProvider> codeFixProviders, int? codeFixIndex, string? codeFixEquivalenceKey, Project project, int numberOfIterations, IVerifier verifier, CancellationToken cancellationToken)
@@ -448,12 +441,11 @@ namespace Microsoft.CodeAnalysis.Testing
                     {
                         anyActions = true;
 
-                        var fixedProject = await ApplyFixAsync(project, actionToApply, cancellationToken).ConfigureAwait(false);
+                        var fixedProject = await ApplyCodeActionAsync(project, actionToApply, verifier, cancellationToken).ConfigureAwait(false);
                         if (fixedProject != project)
                         {
                             done = false;
-
-                            project = await RecreateProjectDocumentsAsync(fixedProject, verifier, cancellationToken).ConfigureAwait(false);
+                            project = fixedProject;
                             break;
                         }
                     }
@@ -486,105 +478,6 @@ namespace Microsoft.CodeAnalysis.Testing
             }
 
             return (project, null);
-        }
-
-        private static CodeAction? TryGetCodeActionToApply(List<CodeAction> actions, int? codeFixIndex, string? codeFixEquivalenceKey, IVerifier verifier)
-        {
-            if (codeFixIndex.HasValue && codeFixEquivalenceKey != null)
-            {
-                if (actions.Count <= codeFixIndex)
-                {
-                    return null;
-                }
-
-                verifier.Equal(
-                    codeFixEquivalenceKey,
-                    actions[codeFixIndex.Value].EquivalenceKey,
-                    "The code action equivalence key and index must be consistent when both are specified.");
-
-                return actions[codeFixIndex.Value];
-            }
-            else if (codeFixEquivalenceKey != null)
-            {
-                return actions.Find(x => x.EquivalenceKey == codeFixEquivalenceKey);
-            }
-            else if (actions.Count > (codeFixIndex ?? 0))
-            {
-                return actions[codeFixIndex ?? 0];
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Implements a workaround for issue #936, force re-parsing to get the same sort of syntax tree as the original document.
-        /// </summary>
-        /// <param name="project">The project to update.</param>
-        /// <param name="verifier">The verifier to use for test assertions.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
-        /// <returns>The updated <see cref="Project"/>.</returns>
-        private async Task<Project> RecreateProjectDocumentsAsync(Project project, IVerifier verifier, CancellationToken cancellationToken)
-        {
-            foreach (var documentId in project.DocumentIds)
-            {
-                var document = project.GetDocument(documentId);
-                var initialTree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
-                document = await RecreateDocumentAsync(document, cancellationToken).ConfigureAwait(false);
-                var recreatedTree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
-                if (CodeFixValidationMode != CodeFixValidationMode.None)
-                {
-                    try
-                    {
-                        // We expect the tree produced by the code fix (initialTree) to match the form of the tree produced
-                        // by the compiler for the same text (recreatedTree).
-                        TreeEqualityVisitor.AssertNodesEqual(
-                            verifier,
-                            SyntaxKindType,
-                            await recreatedTree.GetRootAsync(cancellationToken).ConfigureAwait(false),
-                            await initialTree.GetRootAsync(cancellationToken).ConfigureAwait(false),
-                            checkTrivia: CodeFixValidationMode == CodeFixValidationMode.Full);
-                    }
-                    catch
-                    {
-                        // Try to revalidate the tree with a better message
-                        var renderedInitialTree = TreeToString(await initialTree.GetRootAsync(cancellationToken).ConfigureAwait(false), CodeFixValidationMode);
-                        var renderedRecreatedTree = TreeToString(await recreatedTree.GetRootAsync(cancellationToken).ConfigureAwait(false), CodeFixValidationMode);
-                        verifier.EqualOrDiff(renderedRecreatedTree, renderedInitialTree);
-
-                        // This is not expected to be hit, but it will be hit if the validation failure occurred in a
-                        // portion of the tree not captured by the rendered form from TreeToString.
-                        throw;
-                    }
-                }
-
-                project = document.Project;
-            }
-
-            return project;
-        }
-
-        private static async Task<Document> RecreateDocumentAsync(Document document, CancellationToken cancellationToken)
-        {
-            var newText = await document.GetTextAsync(cancellationToken).ConfigureAwait(false);
-            return document.WithText(SourceText.From(newText.ToString(), newText.Encoding, newText.ChecksumAlgorithm));
-        }
-
-        /// <summary>
-        /// Apply the inputted <see cref="CodeAction"/> to the inputted document.
-        /// Meant to be used to apply code fixes.
-        /// </summary>
-        /// <param name="project">The <see cref="Project"/> to apply the fix on.</param>
-        /// <param name="codeAction">A <see cref="CodeAction"/> that will be applied to the
-        /// <paramref name="project"/>.</param>
-        /// <param name="cancellationToken">The <see cref="CancellationToken"/> that the task will observe.</param>
-        /// <returns>A <see cref="Project"/> with the changes from the <see cref="CodeAction"/>.</returns>
-        private static async Task<Project> ApplyFixAsync(Project project, CodeAction codeAction, CancellationToken cancellationToken)
-        {
-            var operations = await codeAction.GetOperationsAsync(cancellationToken).ConfigureAwait(false);
-            var solution = operations.OfType<ApplyChangesOperation>().Single().ChangedSolution;
-            return solution.GetProject(project.Id);
         }
 
         private Task<(Project project, ExceptionDispatchInfo? iterationCountFailure)> FixAllAnalyzerDiagnosticsInDocumentAsync(ImmutableArray<DiagnosticAnalyzer> analyzers, ImmutableArray<CodeFixProvider> codeFixProviders, int? codeFixIndex, string? codeFixEquivalenceKey, Project project, int numberOfIterations, IVerifier verifier, CancellationToken cancellationToken)
@@ -689,12 +582,11 @@ namespace Microsoft.CodeAnalysis.Testing
                     return (project, null);
                 }
 
-                var fixedProject = await ApplyFixAsync(project, action, cancellationToken).ConfigureAwait(false);
+                var fixedProject = await ApplyCodeActionAsync(project, action, verifier, cancellationToken).ConfigureAwait(false);
                 if (fixedProject != project)
                 {
                     done = false;
-
-                    project = await RecreateProjectDocumentsAsync(fixedProject, verifier, cancellationToken).ConfigureAwait(false);
+                    project = fixedProject;
                 }
             }
             while (!done);
@@ -767,227 +659,6 @@ namespace Microsoft.CodeAnalysis.Testing
             }
 
             return false;
-        }
-
-        private string TreeToString(SyntaxNodeOrToken syntaxNodeOrToken, CodeFixValidationMode validationMode)
-        {
-            var result = new StringBuilder();
-            TreeToString(syntaxNodeOrToken, string.Empty, validationMode, result);
-            return result.ToString();
-        }
-
-        private void TreeToString(SyntaxNodeOrToken syntaxNodeOrToken, string indent, CodeFixValidationMode validationMode, StringBuilder result)
-        {
-            if (syntaxNodeOrToken.IsNode)
-            {
-                result.AppendLine($"{indent}Node({Kind(syntaxNodeOrToken.RawKind)}):");
-
-                var childIndent = indent + "  ";
-                foreach (var child in syntaxNodeOrToken.ChildNodesAndTokens())
-                {
-                    TreeToString(child, childIndent, validationMode, result);
-                }
-            }
-            else
-            {
-                var syntaxToken = syntaxNodeOrToken.AsToken();
-                result.AppendLine($"{indent}Token({Kind(syntaxToken.RawKind)}): {Escape(syntaxToken.Text)}");
-
-                if (validationMode == CodeFixValidationMode.Full)
-                {
-                    var childIndent = indent + "  ";
-                    foreach (var trivia in syntaxToken.LeadingTrivia)
-                    {
-                        if (trivia.HasStructure)
-                        {
-                            result.AppendLine($"{childIndent}Leading({Kind(trivia.RawKind)}):");
-                            TreeToString(trivia.GetStructure(), childIndent + "  ", validationMode, result);
-                        }
-                        else
-                        {
-                            result.AppendLine($"{childIndent}Leading({Kind(trivia.RawKind)}): {Escape(trivia.ToString())}");
-                        }
-                    }
-
-                    foreach (var trivia in syntaxToken.TrailingTrivia)
-                    {
-                        if (trivia.HasStructure)
-                        {
-                            result.AppendLine($"{childIndent}Trailing({Kind(trivia.RawKind)}):");
-                            TreeToString(trivia.GetStructure(), childIndent + "  ", validationMode, result);
-                        }
-                        else
-                        {
-                            result.AppendLine($"{childIndent}Trailing({Kind(trivia.RawKind)}): {Escape(trivia.ToString())}");
-                        }
-                    }
-                }
-            }
-
-            // Local functions
-            string Escape(string text)
-            {
-                return text
-                    .Replace("\\", "\\\\")
-                    .Replace("\t", "\\t")
-                    .Replace("\r", "\\r")
-                    .Replace("\n", "\\n");
-            }
-
-            string Kind(int syntaxKind)
-            {
-                if (SyntaxKindType.GetTypeInfo()?.IsEnum ?? false)
-                {
-                    return Enum.Format(SyntaxKindType, (ushort)syntaxKind, "G");
-                }
-                else
-                {
-                    return syntaxKind.ToString();
-                }
-            }
-        }
-
-        private sealed class SourceFileEqualityComparer : IEqualityComparer<(string filename, SourceText content)>
-        {
-            private SourceFileEqualityComparer()
-            {
-            }
-
-            public static SourceFileEqualityComparer Instance { get; } = new SourceFileEqualityComparer();
-
-            public bool Equals((string filename, SourceText content) x, (string filename, SourceText content) y)
-            {
-                if (x.filename != y.filename)
-                {
-                    return false;
-                }
-
-                if (x.content is null || y.content is null)
-                {
-                    return ReferenceEquals(x, y);
-                }
-
-                return x.content.Encoding == y.content.Encoding
-                    && x.content.ChecksumAlgorithm == y.content.ChecksumAlgorithm
-                    && x.content.ContentEquals(y.content);
-            }
-
-            public int GetHashCode((string filename, SourceText content) obj)
-            {
-                return obj.filename.GetHashCode()
-                    ^ (obj.content?.ToString().GetHashCode() ?? 0);
-            }
-        }
-
-        private class TreeEqualityVisitor
-        {
-            private readonly IVerifier _verifier;
-            private readonly Type _syntaxKindType;
-            private readonly SyntaxNode _expected;
-            private readonly bool _checkTrivia;
-
-            private TreeEqualityVisitor(IVerifier verifier, Type syntaxKindType, SyntaxNode expected, bool checkTrivia)
-            {
-                _verifier = verifier;
-                _syntaxKindType = syntaxKindType;
-                _expected = expected ?? throw new ArgumentNullException(nameof(expected));
-                _checkTrivia = checkTrivia;
-            }
-
-            public void Visit(SyntaxNode node)
-            {
-                AssertSyntaxKindEqual(_expected.RawKind, node.RawKind);
-                AssertChildSyntaxListEqual(_expected.ChildNodesAndTokens(), node.ChildNodesAndTokens(), _checkTrivia);
-            }
-
-            internal static void AssertNodesEqual(IVerifier verifier, Type syntaxKindType, SyntaxNode expected, SyntaxNode actual, bool checkTrivia)
-            {
-                new TreeEqualityVisitor(verifier, syntaxKindType, expected, checkTrivia).Visit(actual);
-            }
-
-            private void AssertNodesEqual(SyntaxNode expected, SyntaxNode actual, bool checkTrivia)
-            {
-                AssertNodesEqual(_verifier, _syntaxKindType, expected, actual, checkTrivia);
-            }
-
-            private void AssertChildSyntaxListEqual(ChildSyntaxList expected, ChildSyntaxList actual, bool checkTrivia)
-            {
-                _verifier.Equal(expected.Count, actual.Count);
-                foreach (var (expectedChild, actualChild) in expected.Zip(actual, (first, second) => (first, second)))
-                {
-                    if (expectedChild.IsToken)
-                    {
-                        _verifier.True(actualChild.IsToken);
-                        AssertTokensEqual(expectedChild.AsToken(), actualChild.AsToken(), checkTrivia);
-                    }
-                    else
-                    {
-                        _verifier.True(actualChild.IsNode);
-                        AssertNodesEqual(expectedChild.AsNode(), actualChild.AsNode(), checkTrivia);
-                    }
-                }
-            }
-
-            private void AssertTokensEqual(SyntaxToken expected, SyntaxToken actual, bool checkTrivia)
-            {
-                AssertTriviaListEqual(expected.LeadingTrivia, actual.LeadingTrivia, checkTrivia);
-                AssertSyntaxKindEqual(expected.RawKind, actual.RawKind);
-                _verifier.Equal(expected.Value, actual.Value);
-                _verifier.Equal(expected.Text, actual.Text);
-                _verifier.Equal(expected.ValueText, actual.ValueText);
-                AssertTriviaListEqual(expected.TrailingTrivia, actual.TrailingTrivia, checkTrivia);
-            }
-
-            private void AssertTriviaListEqual(SyntaxTriviaList expected, SyntaxTriviaList actual, bool checkTrivia)
-            {
-                if (!checkTrivia)
-                {
-                    return;
-                }
-
-                for (var i = 0; i < Math.Min(expected.Count, actual.Count); i++)
-                {
-                    AssertTriviaEqual(expected[i], actual[i], checkTrivia);
-                }
-
-                _verifier.Equal(expected.Count, actual.Count);
-            }
-
-            private void AssertTriviaEqual(SyntaxTrivia expected, SyntaxTrivia actual, bool checkTrivia)
-            {
-                if (!checkTrivia)
-                {
-                    return;
-                }
-
-                AssertSyntaxKindEqual(expected.RawKind, actual.RawKind);
-                _verifier.Equal(expected.HasStructure, actual.HasStructure);
-                _verifier.Equal(expected.IsDirective, actual.IsDirective);
-                _verifier.Equal(expected.GetAnnotations(), actual.GetAnnotations());
-                if (expected.HasStructure)
-                {
-                    AssertNodesEqual(expected.GetStructure(), actual.GetStructure(), checkTrivia);
-                }
-            }
-
-            private void AssertSyntaxKindEqual(int expected, int actual)
-            {
-                if (expected == actual)
-                {
-                    return;
-                }
-
-                if (_syntaxKindType.GetTypeInfo()?.IsEnum ?? false)
-                {
-                    _verifier.Equal(
-                        Enum.Format(_syntaxKindType, (ushort)expected, "G"),
-                        Enum.Format(_syntaxKindType, (ushort)actual, "G"));
-                }
-                else
-                {
-                    _verifier.Equal(expected, actual);
-                }
-            }
         }
     }
 }
